Remove GenericError and refactor
